+++ pyproject.toml
@@ -36,5 +36,5 @@ Homepage = "https://github.com/sig-gis/rlcms/"
 composite = "rlcms.cli.composite:main"
 sample_pts = "rlcms.cli.sample_pts:main"
 train_test = "rlcms.cli.train_test:main"
-primitives = "rlcms.cli.RFprimitives:main"
-generate_LC = "rlcms.cli.generate_LC:main"
+primitives = "rlcms.cli.primitives:main"
+landcover = "rlcms.cli.landcover:main"
